requirements.txt
pandas
numpy
streamlit
geopandas
plotly
python-dotenv